Tweak CSS
* Minor tweaks to layout

Now uses CSS instead of images, uses `max-width` instead of `width`, and changed max width to 920px from 800px (which is a nicer width IMO). :)

* Use max-width for banner logo

responsive~

* Increase size of copyright footer font to 12px

* Added white border

* Padding adjustments

this should be the last one. let's hope i don't find any more

* Added viewport meta tag

* Fixed padding on menu

* update style.css

div #LogoChris {
    max-width: 800px;

would be 600px if qb_banner.png were cropped to be 600px. if only :(

this change makes it so that the height of the div containing the banner stays the same regardless of the size of the browser which changes the height of the banner.

* update header.inc.php

this change makes it so that the height of the div containing the banner stays the same regardless of the size of the browser which changes the height of the banner.

* Remove unused images

* Sort CSS

Co-authored-by: Chocobo1 <Chocobo1@users.noreply.github.com>